create utility function getCanvassers to reuse logic, implement new Person icon and canvassers number in CanvassAssignmentListItem and CanvassAssignmentOverviewlistItem
