+++ src/features/campaigns/components/ActivitiesOverview/items/CanvassAssignmentOverviewListItem.tsx
@@ -1,9 +1,12 @@
 import { FC } from 'react';
-import { Map } from '@mui/icons-material';
+import { Map, Person } from '@mui/icons-material';
 
 import { CanvassAssignmentActivity } from 'features/campaigns/types';
 import getStatusColor from 'features/campaigns/utils/getStatusColor';
 import OverviewListItem from './OverviewListItem';
+import useCanvassSessions from 'features/canvassAssignments/hooks/useCanvassSessions';
+import getCanvassers from 'features/canvassAssignments/utils/getCanvassers';
+import { useNumericRouteParams } from 'core/hooks';
 
 type Props = {
   activity: CanvassAssignmentActivity;
@@ -15,18 +18,26 @@ const CanvassAssignmentOverviewListItem: FC<Props> = ({
   focusDate,
 }) => {
   const assignment = activity.data;
+  const { orgId } = useNumericRouteParams();
+
+  const allSessions = useCanvassSessions(orgId, assignment.id).data || [];
+  const sessions = allSessions.filter(
+    (session) => session.assignment.id === assignment.id
+  );
+
+  const canvassers = getCanvassers(sessions);
 
   return (
     <OverviewListItem
       color={getStatusColor(activity.visibleFrom, activity.visibleUntil)}
       endDate={activity.visibleUntil}
-      endNumber={''}
+      endNumber={canvassers.length}
       focusDate={focusDate}
       href={`/organize/${assignment.organization.id}/projects/${
         assignment.campaign?.id ?? 'standalone'
       }/canvassassignments/${assignment.id}`}
       PrimaryIcon={Map}
-      SecondaryIcon={null}
+      SecondaryIcon={Person}
       startDate={activity.visibleFrom}
       statusBar={null}
       title={assignment.title || ''}
